+++ modules-services/service-templates/src/main/java/org/openprovenance/prov/service/RecordsProcessor.java
@@ -1,7 +1,0 @@
-package org.openprovenance.prov.service;
-
-import java.util.List;
-
-public interface RecordsProcessor<T> {
-    List<T> process(List<Object[]> ll);
-}