+++ RELEASES.md
@@ -13,6 +13,8 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template-compiler: documentation for composite templates
 * prov-template-compiler: forms for composite templates (work in progress)
 * prov-template-compiler: composite templates conversion to csv
+* service-templates: form interface to support composite templates (csv or json)
+
 
 
 `2.0.5`